Update OptionTree plugin to fix PHP 7 compatibility issues.
The OptionTree plugin was outdated and caused PHP notices to be displayed on the page, due to the upgrade from PHP 5 to PHP 7. I've removed the versioned code from the theme (`option-tree` directory) and replaced it with the latest version of the plugin which is now pulled in as a composer package.