Move footer.php and header.php to theme root
WP expects to have a footer and header .php file in the root directory.
To improve the theme's compatibility with our other WP sites and future
plugins I've made this change. I redirected the legacy footer and header
php files rather then delete them, as it seemed a lot of theme files
were pointing to them.