Modify enqueue and Google font script
The theme CSS scripts were loading after the plugin, as they were set to
a very high value 100 ...meaning that any plugin installed, unless over
this value, would get overwriten by the theme. I also removed the Google
font script. It was loading last on the site causing issues with the
site's CSS by bypassing the enqueue process and second, it was not clear
that it was even having an effect on the font styling? It was also
causing the load time to increase.