Update composer file to pull in cookie banner variant B
I had previously added variant B to the composer.json file, but I didn't run composer update, and so it didn't go into the lock file. This meant that when it went up onto dev, the banner wasn't pulled in, as it was looking to the lock file. So this updates the lock file to make sure the banner does get pulled in.